+++ src/components/index.ts
@@ -1,2 +1,3 @@
 export * from './ui';
 export * from './slot';
+export * from './input-group';
