+++ src/components/ui/index.ts
@@ -1,3 +1,2 @@
 export * from './button';
 export * from './input';
-export * from './input-group';
